Extract user CRUD from /test to /user
Move code from auth* to user* where relevant.
