+++ backend/server/controllers/userController.js
@@ -1,0 +1,53 @@
+const userService = require('../services/userService');
+
+const createUser = async (req, res) => {
+  try {
+    const { email, password, confirmPassword } = req.body;
+    await userService.createUser(email, password, confirmPassword);
+    res.json({ message: 'SUCCESS: User created' });
+  } catch (err) {
+    res.status(err?.status || 500)
+       .json({ error: err?.message || err });
+  }
+};
+
+const getUserInfo = async (req, res) => {
+  try {
+    const { id, email } = req.body;
+    const info = await userService.getUserInfo(id, email);
+    res.json({ message: 'SUCCESS', info });
+  } catch (err) {
+    res.status(err?.status || 400)
+       .json({ error: err?.message || err });
+  }
+};
+  
+const updateUser = async (req, res) => {
+  try {
+    const { id, username, password } = req.body;
+    await userService.updateUser(id, username, password);
+    res.json({ message: "SUCCESS: User info updated" });
+  } catch (err) {
+    res.status(err?.status || 400)
+       .json({ error: err?.message || err });
+  }
+};
+  
+const deleteUser = async (req, res) => {
+  try {
+    const { id } = req.body;
+    await userService.deleteUser(id);
+    res.json({ message: 'SUCCESS: User deleted' });
+  } catch (err) {
+    res.status(err?.status || 400)
+       .json({ error: err?.message || err });
+  }
+};
+
+module.exports = {
+  createUser,
+  getUserInfo,
+  updateUser,
+  deleteUser, 
+};
+
