--- backend/server/routes/test.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const authController = require('../controllers/authController.js');
-
-const router = express.Router();
-
-router.post('/getUserInfo', authController.getUserInfo);
-router.post('/update', authController.updateUser);
-router.post('/deregister', authController.deleteUser);
-
-module.exports = router;
